bug deactivating groups from admin page
git-svn-id: svn://svn.newznab.com/nn/trunk@493 a8b8df2f-1e19-4095-b43c-2e6d618ca476

+++ www/admin/ajax_group-edit.php
@@ -7,23 +7,20 @@ require_once(WWW_DIR."/lib/groups.php");
 $admin = new AdminPage;
 $group  = new Groups();
 
-if (isset($_GET['action']))
+if (isset($_GET['action']) && $_GET['action'] == "2")
 {
-	if ($_GET['action'] == "2")
-	{
 		$id     = (int)$_GET['group_id'];
 		$group->delete($id);	
 		print "Group $id deleted.";
-	}
-	else
+}
+else
+{
+	if (isset($_GET['group_id']))
 	{
-		if (isset($_GET['group_id']))
-		{
-			$id     = (int)$_GET['group_id'];
-			$status = isset($_GET['group_status']) ? (int)$_GET['group_status'] : 0;
-			print $group->updateGroupStatus($id, $status);
-		}	
-	}
+		$id     = (int)$_GET['group_id'];
+		$status = isset($_GET['group_status']) ? (int)$_GET['group_status'] : 0;
+		print $group->updateGroupStatus($id, $status);
+	}	
 }
